Quitada la implementación privada de la cámara
Eliminada la clase CameraImplementation. Todas sus funcionalidades ahora están en la clase Camera.

También he quitado un debug en PhysicsManager
