+++ source/Physics/source/PhysicsManager.cpp
@@ -67,9 +67,9 @@ namespace PTSD {
 			{
 				trans = obj->getWorldTransform();
 			}
-			std::string s = "World pos object " + std::to_string(j) + " = " + std::to_string(float(trans.getOrigin().getX())) + ", " +
-				std::to_string(float(trans.getOrigin().getY())) + ", " + std::to_string(float(trans.getOrigin().getZ())) + "\n";
-			LOG(s.c_str(), Trace);
+			//std::string s = "World pos object " + std::to_string(j) + " = " + std::to_string(float(trans.getOrigin().getX())) + ", " +
+			//	std::to_string(float(trans.getOrigin().getY())) + ", " + std::to_string(float(trans.getOrigin().getZ())) + "\n";
+			//LOG(s.c_str(), Trace);
 		}
 	}
