+++ vsprojects/PTSD-Core/PTSD.vcxproj.filters
@@ -84,9 +84,6 @@
     <ClCompile Include="..\..\source\Graphics\source\Camera.cpp">
       <Filter>Graphics\Source</Filter>
     </ClCompile>
-    <ClCompile Include="..\..\source\Graphics\source\CameraImplementation.cpp">
-      <Filter>Graphics\Source</Filter>
-    </ClCompile>
     <ClCompile Include="..\..\source\Graphics\source\GraphicsImplementation.cpp">
       <Filter>Graphics\Source</Filter>
     </ClCompile>
@@ -144,9 +141,6 @@
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="..\..\source\Graphics\include\Camera.h">
-      <Filter>Graphics\Headers</Filter>
-    </ClInclude>
-    <ClInclude Include="..\..\source\Graphics\include\CameraImplementation.h">
       <Filter>Graphics\Headers</Filter>
     </ClInclude>
     <ClInclude Include="..\..\source\Graphics\include\GraphicsImplementation.h">
